pyproject.toml
[tool.poetry]
name = "llm-rag-app"
version = "0.1.0"
description = ""
authors = ["tiziaco <tiziano.iaco@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.1"
ebooklib = "^0.18"
beautifulsoup4 = "^4.12.3"
llama-index-readers-file = "^0.1.12"
llama-index = "^0.10.25"
html2text = "^2024.2.26"
transformers = "^4.39.2"
pgvector = "^0.2.5"
torch = "^2.2.2"
python-dotenv = "^1.0.1"
sentencepiece = "^0.2.0"
langchain = "^0.1.13"
llama-index-vector-stores-postgres = "^0.1.4.post1"
llama-index-embeddings-huggingface = "^0.2.0"
llama-index-llms-llama-cpp = "^0.1.3"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
greenlet = "^3.0.3"
llama-index-llms-ollama = "^0.1.2"
llama-cpp-python = "^0.2.57"
pymupdf = "^1.24.0"
streamlit = "^1.32.2"
llama-index-llms-openllm = "^0.1.3"
llama-index-extractors-entity = "^0.1.2"
spacy = "^3.7.4"
llmsherpa = "^0.1.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
